fix: parser.c: !key_specifed logic in conditions
this adds logic for when the key isn't specified and this also copies
the bytes from the string to a buffer

+++ src/header.h
@@ -1,12 +1,13 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <stdlib.h>
+#include <string.h>
 
 #include "json_data.h"
 
 /* parse.c */
-	int json_parse(const char *file_path, uint16_t num_entries, json_data json_entry[]);
-	uint16_t key_match(Bool *success, const char *key_value, uint16_t num_entries, json_data json_entry[]);
+	int json_parse(const char *file_path, uint8_t num_entries, json_data json_entry[]);
+	uint8_t key_match(Bool *success, const char *key_value, uint8_t num_entries, json_data json_entry[]);
 
 /* compare.c */
 	Bool str_compare(const char *arg, const char *str);
